refactor: use operation.Syntax fallback and split SMA7011 tests
