+++ test/AnalyzerTests/SMA7010_ReflectionAnalyzerTests.cs
@@ -169,55 +169,6 @@ namespace Test
         }
 
         [TestMethod]
-        public async Task SMA7010_Violation_LocalDeclaration()
-        {
-            var test = @"
-using System.Reflection;
-
-namespace Test
-{
-    public class C
-    {
-        public void M()
-        {
-            MethodInfo {|#0:method|} = null;
-            _ = method;
-        }
-    }
-}
-";
-            await VerifyCS.VerifyAnalyzerAsync(test,
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionVariable).WithLocation(0).WithArguments("method", "MethodInfo")
-            );
-        }
-
-        [TestMethod]
-        public async Task SMA7011_Violation_MultiDeclaratorVariableDeclaration()
-        {
-            var test = @"
-using System.Reflection;
-
-namespace Test
-{
-    public class C
-    {
-        static MethodInfo GetIt() => null;
-
-        public void M()
-        {
-            var {|#0:a|}, {|#1:b|} = {|#2:GetIt()|};
-        }
-    }
-}
-";
-            await VerifyCS.VerifyAnalyzerAsync(test,
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionVariable).WithLocation(0).WithArguments("a", "MethodInfo"),
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionVariable).WithLocation(1).WithArguments("b", "MethodInfo"),
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionUsage).WithLocation(2).WithArguments("GetIt", "System.Reflection.MethodInfo")
-            );
-        }
-
-        [TestMethod]
         public async Task SMA7010_Violation_ArgumentType()
         {
             var test = @"
@@ -261,30 +212,6 @@ namespace Test
 ";
             await VerifyCS.VerifyAnalyzerAsync(test,
                 VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionUsage).WithLocation(0).WithArguments("GetMembers", "System.Reflection.MemberInfo")
-            );
-        }
-
-        [TestMethod]
-        public async Task SMA7011_Violation_ForEachVarDeclaration()
-        {
-            var test = @"
-namespace Test
-{
-    public class C
-    {
-        public void M()
-        {
-            foreach (var {|#0:member|} in {|#1:typeof(C).GetMembers()|})
-            {
-                _ = member;
-            }
-        }
-    }
-}
-";
-            await VerifyCS.VerifyAnalyzerAsync(test,
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionVariable).WithLocation(0).WithArguments("member", "MemberInfo"),
-                VerifyCS.Diagnostic(ReflectionAnalyzer.RuleId_SystemReflectionUsage).WithLocation(1).WithArguments("GetMembers", "System.Reflection.MemberInfo")
             );
         }
